agents/content_gen/requirements.txt
flask
flask-socketio
flask-cors
langchain
langchain-community
langchain-openai
langchain-google-genai
python-dotenv
fpdf2
gunicorn
flask-restx
werkzeug